requirements.txt
feedparser==6.0.11
openai-whisper==20231117
requests==2.31.0
torch==2.1.2
elasticsearch==8.13.0
python-dotenv==1.0.1